List of changes:
* Make Modal and PlaceholderBar global components
* Move type definitions to base directory
* Decompose big single file components to multiple sub-components
