+++ app/(root)/_components/Cart/Checkout.tsx
@@ -1,0 +1,3 @@
+export function Checkout() {
+    return (null);
+}
