+++ app/(root)/_components/Cart/cartiteminfo.ts
@@ -1,0 +1,9 @@
+export type CartItemInfo = {
+    id: string;
+    variantId: string | null;
+    quantity: number,
+    name: string;
+    variantName: string | null;
+    price: number;
+    image: string;
+};
