+++ app/(root)/_components/Cart/index.ts
@@ -1,0 +1,3 @@
+import Cart from "./Cart";
+
+export {Cart};
