+++ app/(root)/_components/Navbar/Navbar.tsx
@@ -1,7 +1,7 @@
 "use client"
 
 import Link from "next/link"
-import { LocalCartTransfer, useUser } from "../../_context/user";
+import { LocalCartTransfer, useUser } from "../../../_context/user";
 
 import "@fortawesome/fontawesome-svg-core/styles.css"
 import { config } from "@fortawesome/fontawesome-svg-core";
@@ -10,32 +10,8 @@ config.autoAddCss = false;
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faCartShopping, faUser, faAngleDown, faRightFromBracket } from '@fortawesome/free-solid-svg-icons'
 import { useCallback, useEffect, useRef, useState } from "react";
-import PlaceholderBar from "./PlaceholderBar";
-import { motion, AnimatePresence } from "framer-motion";
-
-const ProfileDropdown = ({ children, close }: { children: React.ReactNode, close: () => void }) => {
-    const wrapperRef = useRef<HTMLDivElement>(null!);
-
-    const onClick = useCallback((e: MouseEvent) => {
-        console.log("Test");
-        if (!wrapperRef.current?.contains(e.target as Node)) { 
-            close(); 
-        }
-    }, [close]);
-
-    useEffect(() => {
-        setTimeout(() => window.addEventListener('click', onClick));
-        return () => window.removeEventListener('click', onClick);
-    }, [onClick]);
-
-    return (
-        <div className="min-w-56 mt-[4.5rem] absolute bg-white top-0 right-0 w-fit border rounded-lg">
-            <div ref={wrapperRef} className="flex flex-col shadow-lg">
-                {children}
-            </div>
-        </div>
-    );
-}
+import PlaceholderBar from "@/components/PlaceholderBar";
+import ProfileDropdown from "./ProfileDropdown";
 
 export default function Navbar() {
     const { setShowLoginModal, user, hasLoggedIn, setCartLoadAction, logout, cartLoaded, getCartItemCount, setCartOpen, loadedUser } = useUser();
