+++ app/(root)/_components/Navbar/index.ts
@@ -1,0 +1,3 @@
+import Navbar from "./Navbar";
+
+export default Navbar;
