+++ app/(root)/shop/page.tsx
@@ -1,5 +1,5 @@
 import { Metadata } from "next";
-import { Product as ProductType } from '../../_types/product'
+import { Product as ProductType } from '../../../types/product'
 import Page from "./_page"
 
 export const metadata: Metadata = {
