+++ app/api/users/update/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import { updateUser } from "../_lib/users";
-import { User } from "../../../_types/user";
+import { User } from "@/types/user";
 
 export const dynamic = 'force-dynamic'
