+++ components/Modal/index.ts
@@ -1,0 +1,5 @@
+import Modal from './Modal'
+
+export type { ModalProps } from './modalprops';
+
+export {Modal};
