+++ components/Modal/modalprops.ts
@@ -1,0 +1,7 @@
+export interface ModalProps {
+    children: React.ReactNode;
+    isModalOpen: boolean,
+    closeModal: () => void;
+    parentStyles?: string;
+    overlayStyles?: string;
+};
