+++ components/PlaceholderBar.tsx
