+++ types/product.ts
